refactor: update cross-chain swap quote functionality
- Implemented adapterQuotes.ts to handle same-chain and fallback quotes for cross-chain swaps.
- Created getPairInfo.ts to fetch and categorize token pairs based on various conditions.
- Developed streamQuotes.ts for streaming quotes from the aggregator API with soft timeout handling.
- Introduced utils.ts for utility functions related to token pricing, categorization, and sorting quotes.
